+++ src/main/java/de/jeff_media/AngelChest/AngelChestCommandUtils.java
@@ -185,7 +185,7 @@ public class AngelChestCommandUtils {
 				unlockCommand="/acunlock " + chestIndex + affectedPlayerParameter;
 			}
 			
-			String text = String.format("[%d] %02d:%02d:%02d §aX:§f %d §aY:§f %d §aZ:§f %d | %s ",
+			String text = String.format("\n[%d] %02d:%02d:%02d §aX:§f %d §aY:§f %d §aZ:§f %d | %s ",
 				chestIndex, hour, min, sec, b.getX(), b.getY(), b.getZ(), b.getWorld().getName()
 			);
 			sendTo.spigot().sendMessage(LinkUtils.getLinks(sendTo, affectedPlayer, text, tpCommand, unlockCommand, fetchCommand, plugin));