+++ src/main/java/de/jeff_media/AngelChest/ConfigUtils.java
@@ -74,6 +74,7 @@ public class ConfigUtils {
 		main.getConfig().addDefault("auto-respawn-delay",10);
 		main.getConfig().addDefault("play-can-skip-auto-respawn",false);
 		main.getConfig().addDefault("use-slimefun",true);
+		main.getConfig().addDefault("show-links-on-separate-line",true);
 		main.getConfig().addDefault("full-xp", false); // Currently not in config because there is no way to get players XP
 		main.disabledWorlds = (ArrayList<String>) main.getConfig().getStringList("disabled-worlds");
 		main.disabledRegions = (ArrayList<String>) main.getConfig().getStringList("disabled-worldguard-regions");